extracting the module checking function to utils module
