+++ lib/commands/new.js
@@ -8,7 +8,7 @@ var logger = require('../logger');
 var prompt = require('readline-sync').question;
 var format = require('../formatters');
 var setup = require('./setup');
-var child_process = require('child_process');
+var utils = require('../utils');
 copy.limit = 16;
 
 module.exports = function () {
@@ -20,26 +20,8 @@ module.exports = function () {
 
     var appFolderName = format.toFolderName(appName);
 
-    try {
-        console.log('Checking for bower installation...');
-        child_process.execSync('npm list -g bower');
-
-    } catch (e) {
-        logger.warn('bower must be installed globally.');
-        console.log('Would you like to install bower now?');
-
-        var result = prompt('Yes'.green + ' or ' + 'No: '.red).toLowerCase();
-
-        if (result == 'yes') {
-            console.log('Installing bower...');
-            child_process.execSync('npm install -g bower');
-            logger.done();
-        }
-
-        if (result == 'no') {
-            console.log('Please install bower and run kick again.');
-            return;
-        }
+    if (!utils.ensureGlobalModules('bower')) {
+        return;
     }
 
     if (fs.existsSync(appFolderName)) {
